pyproject.toml
[project]
name = "thesis-metropolia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncua>=1.1.5",
]

[tool.poetry]
name = "thesis-metropolia"
version = "0.1.0"
description = ""
authors = ["jensjvh <43621451+jensjvh@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"
config = "^0.5.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
